ArmVirtPkg/PlatformBmLib: provide LoadFile2 interface for initrd loading
Implement the new and more generic way of exposing the initrd data to
the Linux kernel, via the LoadFile2 protocol. This way, there is no
longer a need to provide the complicated simple filesystem protocol
at all, given that we can use LoadFile for the kernel image itself.

Note that this approach is intended to be usable across architectures,
by taking the setup_data QEMU fwcfg blob into account as well, which
is only populated on x86.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>