MdeModulePkg/DxeCore/Image: remove unused function arguments
Image.c defines a CoreLoadImageCommon() function that has some arguments
that are marked OPTIONAL. The function only has a single caller living in
the same file, and it does not pass any arguments for these OPTIONAL
arguments.

So let's simplify the code, and remove these arguments and all the
transitive conditional dependencies on them.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>