BaseTools/Scripts/GccBase.lds: export image base symbol
To help converted ELF images perform self-relocation, export a symbol
'ElfImageBase' that can be used in the code to discover the start of
the image in memory.

Note the use of PROVIDE() - this ensures that the symbol is only emitted
if a reference to it exists in the code being linked, but no definition.
This means the symbol is never emitted in a way that can conflict with
existing code.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>